REFACTOR: Adjust some styling and change the folder structure by adding a pages folder to store the pages
