+++ src/components/Banner.jsx
@@ -34,7 +34,7 @@ const Banner = () => {
                 </div>
 
                 <div className="absolute z-50 flex flex-col items-center lg:block lg:left-0  lg:-translate-x-0  bottom-10 md:bottom-14 lg:bottom-1/2 lg:translate-y-1/2 w-full lg:w-[556px] lg:pr-4">
-                  <h1 className="font-josefin text-3xl md:text-5xl font-bold lg:text-4xl xl:text-5xl mb-4 lg:mb-[52px] text-center lg:text-start">
+                  <h1 className="font-josefin text-3xl md:text-5xl lg:text-4xl 2xl:text-5xl font-bold mb-4 lg:mb-[52px] text-center lg:text-start">
                     Personalized Eyeglass Shopping
                   </h1>
                   <Button icon={true}>Buy Now</Button>
