--- src/components/Home.jsx
@@ -1,21 +1,0 @@
-import Experience from "./Experience";
-import Banner from "./Banner";
-import BuyEasyStep from "./BuyEasyStep";
-import Types from "./Types";
-import Products from "./Products";
-import Footer from "./Footer";
-
-const Home = () => {
-  return (
-    <div>
-      <Banner />
-      <BuyEasyStep />
-      <Types />
-      <Experience />
-      <Products />
-      <Footer />
-    </div>
-  );
-};
-
-export default Home;
