--- src/components/Login.jsx
@@ -1,5 +1,0 @@
-const Login = () => {
-  return <div>Login</div>;
-};
-
-export default Login;
