+++ src/pages/Login.jsx
@@ -1,0 +1,5 @@
+const Login = () => {
+  return <div className="container mx-auto">Login</div>;
+};
+
+export default Login;
