+++ src/routes/routes.jsx
@@ -1,8 +1,8 @@
 import { createBrowserRouter } from "react-router-dom";
 import App from "../App";
-import Home from "../components/Home";
-import Cart from "../components/Cart";
-import Login from "../components/Login";
+import Cart from "../pages/Cart";
+import Login from "../pages/Login";
+import Home from "../pages/Home";
 
 const routes = createBrowserRouter([
   {
